+++ src/providers/highlightEdiProvider.ts
@@ -5,7 +5,8 @@ import { EdiUtils } from "../utils/ediUtils";
 
 export class HighlightEdiProvider implements vscode.DocumentHighlightProvider, IProvidable {
   async provideDocumentHighlights(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Promise<vscode.DocumentHighlight[] | null | undefined> {
-    const { parser } = EdiUtils.getEdiParser(document)!;
+    const { parser } = EdiUtils.getEdiParser(document);
+    if (!parser) return;
     const segments = await parser.parseSegments();
     const realPosition = document.offsetAt(new vscode.Position(position.line, position.character));